requirements.txt
# Core dependencies
langgraph>=0.2.0
langgraph-cli
langgraph-cli[inmem]
langchain>=0.3.0
langchain-openai>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# TensorZero and data processing
tensorzero>=0.7.0
pandas>=2.2.0
numpy>=1.26.0
clickhouse-connect>=0.8.0
tenacity>=8.5.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Async support
asyncio>=3.4.3
aiofiles>=24.1.0
tqdm>=4.67.0

# Visualization and reporting
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0

# Testing (optional)
pytest>=8.3.0
pytest-asyncio>=0.25.0

# OpenAI
openai>=1.57.0

# Development tools
black>=24.10.0
ruff>=0.8.0
mypy>=1.13.0